pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "autoformalism-with-llms"
dynamic = ["version"]
description = ''
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = []
authors = [
  { name = "Mike Vaiana", email = "mikevaiana@gmail.com" },
]
classifiers = [
  "Development Status :: 4 - Beta",
  "Programming Language :: Python",
  "Programming Language :: Python :: 3.8",
  "Programming Language :: Python :: 3.9",
  "Programming Language :: Python :: 3.10",
  "Programming Language :: Python :: 3.11",
  "Programming Language :: Python :: 3.12",
  "Programming Language :: Python :: Implementation :: CPython",
  "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = ["pydantic-settings", "python-dotenv", "gitpython", "requests", "tqdm"]

[project.urls]
Documentation = "https://github.com/unknown/autoformalism-with-llms#readme"
Issues = "https://github.com/unknown/autoformalism-with-llms/issues"
Source = "https://github.com/unknown/autoformalism-with-llms"

[tool.hatch.version]
path = "src/autoformalism_with_llms/__about__.py"

[tool.hatch.envs.default]
dependencies = [
  "coverage[toml]>=6.5",
  "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
  "- coverage combine",
  "coverage report",
]
cov = [
  "test-cov",
  "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
  "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/autoformalism_with_llms tests}"

[tool.coverage.run]
source_pkgs = ["autoformalism_with_llms", "tests"]
branch = true
parallel = true
omit = [
  "src/autoformalism_with_llms/__about__.py",
]

[tool.coverage.paths]
autoformalism_with_llms = ["src/autoformalism_with_llms", "*/autoformalism-with-llms/src/autoformalism_with_llms"]
tests = ["tests", "*/autoformalism-with-llms/tests"]

[tool.coverage.report]
exclude_lines = [
  "no cov",
  "if __name__ == .__main__.:",
  "if TYPE_CHECKING:",
]
